Update Slow_Brain.py
Split into 2 modes. User input mode spits out an answer faster and doesnt log anything. Generated mode tests exhaustively and logs the time as well as a list of divisors to update the probability if we build it in.

Time logging in generated mode is intentionally slowed down for comparison with our slow FB...